+++ src/components/Badge/index.tsx
@@ -7,13 +7,20 @@ const variantClassnameMap = {
   dot: 'pui-badge-dot'
 };
 
+const placementClassnameMap = {
+  ['top-right']: 'pui-badge-placement-top-right',
+  ['bottom-right']: 'pui-badge-placement-bottom-right',
+  ['top-left']: 'pui-badge-placement-top-left',
+  ['bottom-left']: 'pui-badge-placement-bottom-left'
+};
+
 export type BadgeProps = {
   /**
-   * The badge display number
+   * The badge display number or text
    */
-  content?: number;
+  content?: number | string;
   /**
-   * The maximum number until the badge will add a "+" sign
+   * The maximum number until the badge will add a "+" sign, only taken into consideration if `content` is of type number
    *
    * @default 9
    */
@@ -28,16 +35,23 @@ export type BadgeProps = {
    * @default default
    */
   variant?: keyof typeof variantClassnameMap;
+  /**
+   * The badge placement
+   *
+   * @default top-right
+   */
+  placement?: keyof typeof placementClassnameMap;
 };
 
 /**
  * A badge
  *
  * @param props the component props
- * @param props.content The badge display number
- * @param props.maxValue The maximum number until the badge will add a "+" sign
+ * @param props.content The badge display number or text
+ * @param props.maxValue The maximum number until the badge will add a "+" sign, only taken into consideration if `content` is of type number
  * @param props.pulse Whether or not the badge should have a pulsing effect
  * @param props.variant The badge variant
+ * @param props.placement The bagge placement
  * @param props.children the provided children
  */
 export const Badge: React.FC<BadgeProps> = ({
@@ -45,18 +59,22 @@ export const Badge: React.FC<BadgeProps> = ({
   maxValue = 9,
   pulse,
   variant = 'default',
+  placement = 'top-right',
   children
 }) => {
-  const content = useMemo(() => (propsContent && propsContent > maxValue ? `${maxValue}+` : propsContent), [
-    propsContent,
-    maxValue
-  ]);
+  const content = useMemo(
+    () =>
+      typeof propsContent === 'number' ? (propsContent > maxValue ? `${maxValue}+` : `${propsContent}`) : propsContent,
+    [propsContent, maxValue]
+  );
 
   return (
     <div className="pui-badge">
       {children}
       {content && (
-        <span className={classnames(variantClassnameMap[variant], { pulse })}>{variant !== 'dot' && content}</span>
+        <span className={classnames(variantClassnameMap[variant], placementClassnameMap[placement], { pulse })}>
+          {variant !== 'dot' && content}
+        </span>
       )}
     </div>
   );